[charts][docs] Improve Charts export docs
