+++ docs/data/charts/export/ExportChartAsImage.js
@@ -15,13 +15,7 @@ function ExportParamsSelector({ apiRef }) {
   const quality = Math.max(0, Math.min(1, Number.parseFloat(rawQuality)));
 
   return (
-    <Stack
-      direction="row"
-      justifyContent="space-between"
-      flexWrap="wrap"
-      gap={2}
-      sx={{ width: '100%' }}
-    >
+    <Stack justifyContent="space-between" gap={2} sx={{ width: '100%' }}>
       <FormControl fullWidth>
         <FormLabel id="image-format-radio-buttons-group-label">
           Image Format
@@ -59,12 +53,14 @@ function ExportParamsSelector({ apiRef }) {
           helperText="Only applicable to lossy formats."
         />
       </FormControl>
-      <Button
-        variant="outlined"
-        onClick={() => apiRef.current?.exportAsImage({ type, quality })}
-      >
-        Export Image
-      </Button>
+      <div>
+        <Button
+          onClick={() => apiRef.current.exportAsImage({ type, quality })}
+          variant="contained"
+        >
+          Export Image
+        </Button>
+      </div>
     </Stack>
   );
 }
