--- docs/data/charts/export/PrintChart.tsx.preview
@@ -1,15 +1,0 @@
-<ScatterChartPro
-  apiRef={apiRef}
-  height={300}
-  series={[
-    {
-      label: 'Series A',
-      data: data.map((v) => ({ x: v.x1, y: v.y1, id: v.id })),
-    },
-    {
-      label: 'Series B',
-      data: data.map((v) => ({ x: v.x1, y: v.y2, id: v.id })),
-    },
-  ]}
-/>
-<Button onClick={() => apiRef.current?.exportAsPrint()}>Print</Button>
